add signInfo.js to container
