+++ src/App.js
@@ -9,7 +9,7 @@ import {SIGN_IMAGES, HOST} from './constants'
 import Messenger from './containers/Messenger'
 import DiscoverPage from './containers/DiscoverPage'
 import ControllerProfilePage from './containers/ControllerProfilePage'
-import SignInfo from './components/SignInfo'
+import SignInfo from './containers/SignInfo'
 import ZodiacFormContainer from './components/ZodiacFormContainer'
 import EditProfile from './containers/EditProfile'
 import './CSS/App.scss'
